ボイボ寮のキャラ個別ページからダウンロードを可能にする
