+++ src/pages/dormitory.tsx
@@ -13,6 +13,8 @@ type DormitoryProps = {
   setShowingHeader: (show: boolean) => void
 }
 
+// FIXME: ハッシュ付きURLに遷移してページ移動したあとブラウザバックするとそのハッシュ位置に戻ってしまうバグを防ぎたい
+// https://github.com/VOICEVOX/voicevox_blog/issues/104
 const Dormitory: React.FC<DormitoryProps> = ({ setShowingHeader }) => {
   const { characterInfos } = useDetailedCharacterInfo()
